requirements.txt
ollama>=0.1.6
streamlit>=1.31.0
faiss-cpu>=1.7.4
chromadb>=0.4.22
python-dotenv>=1.0.0
pypdf>=4.0.1
sentence-transformers>=2.5.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.6.0
beautifulsoup4==4.12.2
requests==2.31.0
tqdm>=4.66.0
numpy>=1.24.0
pyyaml>=6.0.0
python-multipart==0.0.6 